application/src/main/java/com/gsv/codeflix/admin/catalog/application/category/delete/DefaultDeleteCategoryUseCase.java
package com.gsv.codeflix.admin.catalog.application.category.delete;

import com.gsv.codeflix.admin.catalog.domain.category.CategoryGateway;
import com.gsv.codeflix.admin.catalog.domain.category.CategoryID;

public class DefaultDeleteCategoryUseCase extends DeleteCategoryUseCase {

    final CategoryGateway categoryGateway;

    public DefaultDeleteCategoryUseCase(CategoryGateway categoryGateway) {
        this.categoryGateway = categoryGateway;
    }

    @Override
    public void execute(final String id) {
        final var categoryID = CategoryID.from(id);
        categoryGateway.deleteById(categoryID);
    }
}
